+++ griptape/tools/__init__.py
@@ -24,7 +24,7 @@ from .variation_image_generation_client.tool import VariationImageGenerationClie
 from .inpainting_image_generation_client.tool import InpaintingImageGenerationClient
 from .outpainting_image_generation_client.tool import OutpaintingImageGenerationClient
 from .griptape_cloud_knowledge_base_client.tool import GriptapeCloudKnowledgeBaseClient
-from .griptape_cloud_structure_run_client.tool import GriptapeCloudStructureRunClient
+from .structure_run_client.tool import StructureRunClient
 from .image_query_client.tool import ImageQueryClient
 
 __all__ = [
@@ -54,6 +54,6 @@ __all__ = [
     "InpaintingImageGenerationClient",
     "OutpaintingImageGenerationClient",
     "GriptapeCloudKnowledgeBaseClient",
-    "GriptapeCloudStructureRunClient",
+    "StructureRunClient",
     "ImageQueryClient",
 ]